backend/requirements.txt
fastapi[standard]
pydantic
passlib[bcrypt]
python-jose
typing
python-dotenv
pytest
httpx
pyjwt
sqlalchemy
python-multipart

